+++ transition_scripts/rotate.sh
@@ -13,13 +13,13 @@ HEIGHT=720
 FPS=30
 TRANSITION_DURATION=1
 SCREEN_MODE=4                # 1=CENTER, 2=CROP, 3=SCALE, 4=BLUR
-PHOTO_DURATION=2
+IMAGE_DURATION=2
 BACKGROUND_COLOR="black"
 
 IFS=$'\t\n'                 # REQUIRED TO SUPPORT SPACES IN FILE NAMES
 
 # FILE OPTIONS
-# FILES=`find ../media/*.jpg | sort -r`                 # USE ALL IMAGES UNDER THE media FOLDER SORTED
+# FILES=`find ../media/*.jpg | sort -r`             # USE ALL IMAGES UNDER THE media FOLDER SORTED
 # FILES=('../media/1.jpg' '../media/2.jpg')         # USE ONLY THESE IMAGE FILES
 FILES=`find ../media/*.jpg`                         # USE ALL IMAGES UNDER THE media FOLDER
 
@@ -28,21 +28,21 @@ FILES=`find ../media/*.jpg`                         # USE ALL IMAGES UNDER THE m
 ############################
 
 # CALCULATE LENGTH MANUALLY
-let PHOTOS_COUNT=0
-for photo in ${FILES[@]}; do (( PHOTOS_COUNT+=1 )); done
+let IMAGE_COUNT=0
+for IMAGE in ${FILES[@]}; do (( IMAGE_COUNT+=1 )); done
 
-if [[ ${PHOTOS_COUNT} -lt 2 ]]; then
-    echo "Error: photos folder should contain at least two photos"
+if [[ ${IMAGE_COUNT} -lt 2 ]]; then
+    echo "Error: media folder should contain at least two images"
     exit 1;
 fi
 
 # INTERNAL VARIABLES
 TRANSITION_FRAME_COUNT=$(( TRANSITION_DURATION*FPS ))
-PHOTO_FRAME_COUNT=$(( PHOTO_DURATION*FPS ))
-TOTAL_DURATION=$(( (PHOTO_DURATION+TRANSITION_DURATION)*PHOTOS_COUNT))
+IMAGE_FRAME_COUNT=$(( IMAGE_DURATION*FPS ))
+TOTAL_DURATION=$(( (IMAGE_DURATION+TRANSITION_DURATION)*IMAGE_COUNT))
 TOTAL_FRAME_COUNT=$(( TOTAL_DURATION*FPS ))
 
-echo -e "\nVideo Slideshow Info\n------------------------\nPhoto count: ${PHOTOS_COUNT}\nDimension: ${WIDTH}x${HEIGHT}\nFPS: ${FPS}\nPhoto duration: ${PHOTO_DURATION} s\n\
+echo -e "\nVideo Slideshow Info\n------------------------\nImage count: ${IMAGE_COUNT}\nDimension: ${WIDTH}x${HEIGHT}\nFPS: ${FPS}\nImage duration: ${IMAGE_DURATION} s\n\
 Transition duration: ${TRANSITION_DURATION} s\nTotal duration: ${TOTAL_DURATION} s\n"
 
 START_TIME=$SECONDS
@@ -51,8 +51,8 @@ START_TIME=$SECONDS
 FULL_SCRIPT="ffmpeg -y "
 
 # 2. ADD INPUTS
-for photo in ${FILES[@]}; do
-    FULL_SCRIPT+="-loop 1 -i '${photo}' "
+for IMAGE in ${FILES[@]}; do
+    FULL_SCRIPT+="-loop 1 -i '${IMAGE}' "
 done
 
 # 3. ADD BACKGROUND COLOR SCREEN INPUT
@@ -62,14 +62,14 @@ FULL_SCRIPT+="-f lavfi -i color=${BACKGROUND_COLOR}:s=${WIDTH}x${HEIGHT},fps=${F
 FULL_SCRIPT+="-filter_complex \""
 
 # 5. PREPARE INPUTS
-for (( c=0; c<${PHOTOS_COUNT}; c++ ))
+for (( c=0; c<${IMAGE_COUNT}; c++ ))
 do
     case ${SCREEN_MODE} in
         1)
             FULL_SCRIPT+="[${c}:v]setpts=PTS-STARTPTS,scale=w='if(gte(iw/ih,${WIDTH}/${HEIGHT}),min(iw,${WIDTH}),-1)':h='if(gte(iw/ih,${WIDTH}/${HEIGHT}),-1,min(ih,${HEIGHT}))',scale=trunc(iw/2)*2:trunc(ih/2)*2,setsar=sar=1/1,fps=${FPS},format=rgba,split=2[stream$((c+1))pre1][stream$((c+1))pre2];"
         ;;
         2)
-            FULL_SCRIPT+="[${c}:v]setpts=PTS-STARTPTS,scale=w='if(gte(iw/ih,${WIDTH}/${HEIGHT}),-1,${WIDTH})':h='if(gte(iw/ih,${WIDTH}/${HEIGHT}),${HEIGHT},-1)',[${PHOTOS_COUNT}:v]overlay,crop=${WIDTH}:${HEIGHT},setsar=sar=1/1,fps=${FPS},format=rgba,split=2[stream$((c+1))pre1][stream$((c+1))pre2];"
+            FULL_SCRIPT+="[${c}:v]setpts=PTS-STARTPTS,scale=w='if(gte(iw/ih,${WIDTH}/${HEIGHT}),-1,${WIDTH})':h='if(gte(iw/ih,${WIDTH}/${HEIGHT}),${HEIGHT},-1)',[${IMAGE_COUNT}:v]overlay,crop=${WIDTH}:${HEIGHT},setsar=sar=1/1,fps=${FPS},format=rgba,split=2[stream$((c+1))pre1][stream$((c+1))pre2];"
         ;;
         3)
             FULL_SCRIPT+="[${c}:v]setpts=PTS-STARTPTS,scale=${WIDTH}:${HEIGHT},setsar=sar=1/1,fps=${FPS},format=rgba,split=2[stream$((c+1))pre1][stream$((c+1))pre2];"
@@ -83,10 +83,10 @@ do
 done
 
 # 6. APPLY PADDING
-for (( c=1; c<=${PHOTOS_COUNT}; c++ ))
+for (( c=1; c<=${IMAGE_COUNT}; c++ ))
 do
 
-    if [[ ${c} -eq ${PHOTOS_COUNT} ]]; then
+    if [[ ${c} -eq ${IMAGE_COUNT} ]]; then
         SPLIT_COUNT="2"
     else
         SPLIT_COUNT="3"
@@ -96,7 +96,7 @@ do
 
     FULL_SCRIPT+="[stream${c}pre2]pad=width=${WIDTH}:height=${HEIGHT}:x=(${WIDTH}-iw)/2:y=(${HEIGHT}-ih)/2:color=${BACKGROUND_COLOR},trim=duration=${TRANSITION_DURATION},select=lte(n\,${TRANSITION_FRAME_COUNT}),split=${SPLIT_COUNT}[stream${c}out2][stream${c}out3]"
 
-    if [[ ${c} -eq ${PHOTOS_COUNT} ]]; then
+    if [[ ${c} -eq ${IMAGE_COUNT} ]]; then
         FULL_SCRIPT+=";"
     else
         FULL_SCRIPT+="[stream${c}out4];"
@@ -104,10 +104,10 @@ do
 done
 
 # 7. CREATE ROTATION FRAMES
-for (( c=1; c<=${PHOTOS_COUNT}; c++ ))
+for (( c=1; c<=${IMAGE_COUNT}; c++ ))
 do
     if [[ ${c} -eq 1 ]]; then
-        OVERLAY_FRAME="${PHOTOS_COUNT}:v"
+        OVERLAY_FRAME="${IMAGE_COUNT}:v"
     else
         OVERLAY_FRAME="stream$((c-1))out2"
     fi
@@ -116,13 +116,13 @@ do
 done
 
 # 8. BEGIN CONCAT
-for (( c=1; c<${PHOTOS_COUNT}; c++ ))
+for (( c=1; c<${IMAGE_COUNT}; c++ ))
 do
     FULL_SCRIPT+="[stream${c}rotating][stream${c}out3][stream${c}out4]"
 done
 
 # 9. END CONCAT
-FULL_SCRIPT+="[stream${PHOTOS_COUNT}rotating][stream${PHOTOS_COUNT}out2][stream${PHOTOS_COUNT}out3]concat=n=$((3*PHOTOS_COUNT)):v=1:a=0,format=yuv420p[video]\""
+FULL_SCRIPT+="[stream${IMAGE_COUNT}rotating][stream${IMAGE_COUNT}out2][stream${IMAGE_COUNT}out3]concat=n=$((3*IMAGE_COUNT)):v=1:a=0,format=yuv420p[video]\""
 
 # 10. END
 FULL_SCRIPT+=" -map [video] -vsync 2 -async 1 -rc-lookahead 0 -g 0 -profile:v main -level 42 -c:v libx264 -r ${FPS} ../transition_rotate.mp4"